backend/requirements.txt
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.2.0
pymysql==1.1.0
APScheduler==3.10.4
python-dotenv==1.0.0
cryptography==41.0.3
